refactor: move client features
