+++ src/lib/features/client-feature-toggles/tests/client-feature-toggle.e2e.test.ts
@@ -1,12 +1,12 @@
 import supertest from 'supertest';
-import createStores from '../../../test/fixtures/store';
-import getLogger from '../../../test/fixtures/no-logger';
-import getApp from '../../app';
-import { createServices } from '../../services';
-import FeatureController from './feature';
-import { createTestConfig } from '../../../test/config/test-config';
+import createStores from '../../../../test/fixtures/store';
+import getLogger from '../../../../test/fixtures/no-logger';
+import getApp from '../../../app';
+import { createServices } from '../../../services';
+import FeatureController from '../client-feature-toggle.controller';
+import { createTestConfig } from '../../../../test/config/test-config';
 import { secondsToMilliseconds } from 'date-fns';
-import { ClientSpecService } from '../../services/client-spec-service';
+import { ClientSpecService } from '../../../services/client-spec-service';
 
 async function getSetup() {
     const base = `/random${Math.round(Math.random() * 1000)}`;
